login / registerpage splitted and implemented better exception messages
